requirements.txt
Flask
Flask-Cors
python-dotenv
python-docx
PyPDF2
requests
gunicorn
# youtube-transcript-api
# git+https://github.com/openai/whisper.git
# pytube
# torch